Break out assertions into types and macros
