+++ src/evaluators.c
@@ -1,34 +1,7 @@
 #include "evaluators.h"
+#include "output.h"
 #include <stdlib.h>
 #include <stdio.h>
-
-// General purpose converter from string (==) to operator enum (CAUGHT_OP_EQUAL)
-enum caught_operator
-caught_str_to_operator(char *str)
-{
-    int len = sizeof(CAUGHT_OPERATOR_STRS) / sizeof(CAUGHT_OPERATOR_STRS[0]);
-    int i;
-    for (i = 0; i < len; ++i)
-    {
-        if (strcmp(str, CAUGHT_OPERATOR_STRS[i]) == 0)
-        {
-            return i;
-        }
-    }
-    return -1;
-}
-
-// General purpose converter from enum (CAUGHT_OP_EQUAL) to operator string (==)
-const char *caught_operator_to_str(enum caught_operator operator)
-{
-    return CAUGHT_OPERATOR_STRS[operator];
-}
-
-// General purpose converter from enum (CAUGHT_OP_EQUAL) to a to be statement (to be, to not be, etc.)
-const char *caught_operator_to_to_be_statement(enum caught_operator operator)
-{
-    return CAUGHT_OPERATOR_TO_BES[operator];
-}
 
 // Evaluators take in a left hand size, operator, and right hand side
 // they then evaluate the result of that expression
@@ -85,8 +58,8 @@ bool caught_internal_evaluator_str(char *lhs, enum caught_operator operator, cha
     case CAUGHT_OP_NOT_EQUAL:
         return (null_exists && lhs != rhs) || (!null_exists && strcmp(lhs, rhs) != 0);
     default:
-        fprintf(stderr, "Cannot compare strings with %s, only == and != are supported!", caught_operator_to_str(operator));
-        exit(1);
+        caught_output_errorf("Cannot compare strings with %s, only == and != are supported!", caught_operator_to_str(operator));
+        return false;
     }
 }
 bool caught_internal_evaluator_str_ptr(char **lhs, enum caught_operator operator, char ** rhs)
@@ -99,8 +72,7 @@ bool caught_internal_evaluator_exit_status(caught_internal_process_status lhs, e
 {
     if (operator!= CAUGHT_OP_EQUAL)
     {
-        fprintf(stderr, "Cannot compare exit statuses with %s, only == is supported!", caught_operator_to_str(operator));
-        exit(1);
+        caught_output_errorf("Cannot compare exit statuses with %s, only == is supported!", caught_operator_to_str(operator));
     }
 
     return lhs.type == rhs.type && lhs.status == rhs.status;
