+++ src/evaluators.h
@@ -1,42 +1,10 @@
 #include <stdbool.h>
 #include <string.h>
 #include "fork.h"
+#include "assertion-result.h"
 
 #ifndef CAUGHT_EVALUATORS
 #define CAUGHT_EVALUATORS
-
-enum caught_operator
-{
-    CAUGHT_OP_EQUAL,
-    CAUGHT_OP_NOT_EQUAL,
-    CAUGHT_OP_LESS_THAN,
-    CAUGHT_OP_GREATER_THAN,
-    CAUGHT_OP_LESS_THAN_EQ,
-    CAUGHT_OP_GREATER_THAN_EQ,
-};
-
-static char *CAUGHT_OPERATOR_STRS[] = {
-    "==",
-    "!=",
-    "<",
-    ">",
-    "<=",
-    ">=",
-};
-
-static char *CAUGHT_OPERATOR_TO_BES[] = {
-    "to be",
-    "to not be",
-    "to be less than",
-    "to be greater than",
-    "to be <= to",
-    "to be >= to",
-};
-
-enum caught_operator
-caught_str_to_operator(char *str);
-const char *caught_operator_to_str(enum caught_operator operator);
-const char *caught_operator_to_to_be_statement(enum caught_operator operator);
 
 bool caught_internal_evaluator_ptr(void *lhs, enum caught_operator operator, void * rhs);
 bool caught_internal_evaluator_ptr_ptr(void **lhs, enum caught_operator operator, void ** rhs);
