+++ src/output.h
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 #include <string.h>
-#include "assertions.h"
+#include "assertion-result.h"
 #ifndef CAUGHT_OUTPUT
 #define CAUGHT_OUTPUT
